Add the initial bsp for FRDM-KE02Z
The bsp is portinged from the offical bsp for FRDM-KL25Z.
For the SRAM size of KE02Z is just 4K Bytes, so some OS features, e.g.
OS_CFG_PRIO_MAX, need to be tailored. The uC/Shell would almost drained
the SRAM, so it is not recommended to use uC/Shell in the application
usage for this board. Although it is enabled default in this first demo.

Signed-off-by: Ian Peace <ian.peace@outlook.com>